lambdas/pdf-converter/requirements.txt
PyMuPDF==1.23.8
Pillow==10.1.0
boto3==1.34.0
